+++ pkg/analyzer/lib/src/summary2/ast_binary_tag.dart
@@ -121,6 +121,7 @@ enum AstNodeTag {
   ImportPrefixedNameExpression,
   ImplicitFunctionInstantiation,
   ImplicitCallTearOff,
+  FunctionInstantiation,
 }
 
 enum DirectiveUriKind {